fix: 修复数据库统计查询类型转换错误 🐛
问题：
- loadDatabaseInfo() 中查询 COUNT(*) 结果尝试转换为 Int 失败
- DatabaseManager.executeQuery() 对 SQLITE_INTEGER 返回 Int64
- 导致 usageRecordsCount 和 shortcutsCount 始终为 0

修复：
- 将 as? Int 改为 as? Int64，然后转换为 Int(count)
- 修复使用记录数显示
- 修复快捷键数显示

数据验证：
- 数据库实际包含 2918 条使用记录（通过 sqlite3 命令验证）
- 修复后 UI 将正确显示数据

+++ Keymap/UI/Views/Settings/SettingsWindow.swift
@@ -1637,15 +1637,15 @@ class SettingsViewModel: ObservableObject {
         // 获取使用记录数
         let usageSql = "SELECT COUNT(*) as count FROM usage_records"
         let usageResults = databaseManager.executeQuery(usageSql)
-        if let first = usageResults.first, let count = first["count"] as? Int {
-            usageRecordsCount = count
+        if let first = usageResults.first, let count = first["count"] as? Int64 {
+            usageRecordsCount = Int(count)
         }
 
         // 获取快捷键数
         let shortcutsSql = "SELECT COUNT(*) as count FROM shortcuts"
         let shortcutsResults = databaseManager.executeQuery(shortcutsSql)
-        if let first = shortcutsResults.first, let count = first["count"] as? Int {
-            shortcutsCount = count
+        if let first = shortcutsResults.first, let count = first["count"] as? Int64 {
+            shortcutsCount = Int(count)
         }
     }
